Reorganise field management code.
